[WIP] storage/files: remove "FileStorageInterface"
