+++ internal/storage/files/types.go
@@ -1,10 +1,9 @@
 package files
 
 import (
-	"io"
-	"mime/multipart"
 	"time"
 
+	"github.com/tags-drive/core/internal/storage/files/aggregation"
 	"github.com/tags-drive/core/internal/storage/files/extensions"
 )
 
@@ -35,55 +34,6 @@ type GetFilesConfig struct {
 	Filter   FilterFilesFunction // can be nil
 }
 
-// FileStorageInterface provides methods for interactions with files
-type FileStorageInterface interface {
-	// Start starts all background services
-	StartBackgroundServices()
-
-	// Get returns all "good" sorted files
-	//
-	// If cnf.Expr isn't valid, Get returns ErrBadExpessionSyntax
-	Get(cnf GetFilesConfig) ([]File, error)
-	// GetFile returns a file with passed id
-	GetFile(id int) (File, error)
-	// CheckFile checks if file with passed id exists
-	CheckFile(id int) bool
-	// GetFiles returns files with passed ids
-	GetFiles(ids ...int) []File
-	// GetRecent returns the last uploaded files
-	GetRecent(number int) []File
-	// ArchiveFiles archives passed files and returns io.Reader with archive
-	Archive(fileIDs []int) (io.Reader, error)
-
-	// UploadFile uploads a new file
-	Upload(file *multipart.FileHeader, tags []int) error
-
-	// Rename renames a file
-	Rename(fileID int, newName string) (updatedFile File, err error)
-	// ChangeTags changes the tags
-	ChangeTags(fileID int, tags []int) (updatedFile File, err error)
-	// ChangeDescription changes the description
-	ChangeDescription(fileID int, newDescription string) (updatedFile File, err error)
-
-	// Delete "move" a file into Trash
-	Delete(fileID int) error
-	// DeleteForce deletes file from storage and from disk
-	DeleteForce(fileID int) error
-	// Recover "removes" file from Trash
-	Recover(fileID int)
-
-	// AddTagsToFiles adds a tag to files
-	AddTagsToFiles(filesIDs, tagsIDs []int)
-	// RemoveTagsFromFiles
-	RemoveTagsFromFiles(filesIDs, tagsIDs []int)
-
-	// DeleteTagFromFiles deletes a tag from files
-	DeleteTagFromFiles(tagID int)
-
-	// Shutdown gracefully shutdown FileStorage
-	Shutdown() error
-}
-
 // File contains the information about a file
 type File struct {
 	ID       int            `json:"id"`
@@ -111,3 +61,57 @@ const (
 	SortBySizeAsc
 	SortBySizeDecs
 )
+
+// storage is an internal storage for files metadata
+type internalStorage interface {
+	init() error
+
+	// getFile returns a file with passed filename
+	getFile(id int) (File, error)
+
+	checkFile(id int) bool
+
+	// getFiles returns files
+	//     expr - parsed logical expression
+	//     search - string, which filename has to contain (lower case)
+	//     isRegexp - is expr a regular expression (if it is true, expr must be valid regular expression)
+	getFiles(expr aggregation.LogicalExpr, search string, isRegexp bool) (files []File)
+
+	getFilesWithIDs(ids ...int) []File
+
+	// add adds a file
+	addFile(filename string, fileType extensions.Ext, tags []int, size int64, addTime time.Time) (id int)
+
+	// renameFile renames a file
+	renameFile(id int, newName string) (File, error)
+
+	// updateFileTags updates tags of a file
+	updateFileTags(id int, changedTagsID []int) (File, error)
+
+	// updateFileDescription update description of a file
+	updateFileDescription(id int, newDesc string) (File, error)
+
+	// deleteFile marks file deleted and sets TimeToDelete
+	// File can't be deleted several times (function should return ErrFileDeletedAgain)
+	deleteFile(id int) error
+
+	// deleteFileForce deletes file
+	deleteFileForce(id int) error
+
+	// recover removes file from Trash
+	recover(id int)
+
+	// addTagsToFiles adds a tag to files
+	addTagsToFiles(filesIDs, tagsID []int)
+
+	// removeTagsFromFiles removes tags from selected files
+	removeTagsFromFiles(filesIDs, tagsID []int)
+
+	// deleteTagFromFiles deletes a tag
+	deleteTagFromFiles(tagID int)
+
+	// getExpiredDeletedFiles returns names of files with expired TimeToDelete
+	getExpiredDeletedFiles() []int
+
+	shutdown() error
+}
